refactor: Extract components to separate files
Just organize this slightly better
